requirements.txt
agno>=2.5.0
openai>=1.0.0
anthropic>=0.40.0
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
